Add Weak Link security checks for front and back end

+++ wildlifecompliance/components/main/views.py
@@ -1,7 +1,7 @@
 import traceback
 from django.core.exceptions import ValidationError
 from django.db import transaction
-from rest_framework import serializers, views
+from rest_framework import serializers, views, status
 from rest_framework.response import Response
 from rest_framework.renderers import JSONRenderer
 from wildlifecompliance.components.main.utils import (
@@ -56,28 +56,33 @@ class CreateWeakLinkView(views.APIView):
                 first_object_id = request.data.get('first_object_id')
                 second_content_type_str = request.data.get('second_content_type')
                 second_object_id = request.data.get('second_object_id')
+                can_user_action = request.data.get('can_user_action')
                 
-                # transform request data to create new Weak Links obj
-                second_object_id_int = int(second_object_id)
-                first_content_type = ContentType.objects.get(
-                        app_label='wildlifecompliance', 
-                        model=first_content_type_str)
-                second_content_type = ContentType.objects.get(
-                        app_label='wildlifecompliance', 
-                        model=second_content_type_str)
-                
-                weak_link_instance, created = WeakLinks.objects.get_or_create(
-                        first_content_type_id = first_content_type.id,
-                        first_object_id = first_object_id,
-                        second_content_type_id = second_content_type.id,
-                        second_object_id = second_object_id_int
-                        )
-                # derive parent (calling) object instance from weak_link_instance
-                calling_instance = weak_link_instance.first_content_type.model_class().objects.get(id=first_object_id)
+                if can_user_action:
+                    # transform request data to create new Weak Links obj
+                    second_object_id_int = int(second_object_id)
+                    first_content_type = ContentType.objects.get(
+                            app_label='wildlifecompliance', 
+                            model=first_content_type_str)
+                    second_content_type = ContentType.objects.get(
+                            app_label='wildlifecompliance', 
+                            model=second_content_type_str)
+                    
+                    weak_link_instance, created = WeakLinks.objects.get_or_create(
+                            first_content_type_id = first_content_type.id,
+                            first_object_id = first_object_id,
+                            second_content_type_id = second_content_type.id,
+                            second_object_id = second_object_id_int
+                            )
+                    # derive parent (calling) object instance from weak_link_instance
+                    calling_instance = weak_link_instance.first_content_type.model_class().objects.get(id=first_object_id)
 
-                # get related items of calling_instance
-                related_items = get_related_items(calling_instance)
-                return Response(related_items)
+                    # get related items of calling_instance
+                    related_items = get_related_items(calling_instance)
+                    return Response(related_items)
+                else:
+                    content = {'message': 'User does not have permission to perform this action'}
+                    return Response(content, status=status.HTTP_403_FORBIDDEN)
 
         except Exception as e:
             print(traceback.print_exc())
@@ -94,43 +99,48 @@ class RemoveWeakLinkView(views.APIView):
                 first_object_id = request.data.get('first_object_id')
                 second_content_type_str = request.data.get('second_content_type')
                 second_object_id = request.data.get('second_object_id')
-                
-                # transform request data to search for Weak Link obj to delete
-                second_object_id_int = int(second_object_id)
-                first_content_type = ContentType.objects.get(
-                        app_label='wildlifecompliance', 
-                        model=first_content_type_str)
-                second_content_type = ContentType.objects.get(
-                        app_label='wildlifecompliance', 
-                        model=second_content_type_str)
+                can_user_action = request.data.get('can_user_action')
 
-                # If weak link obj not found, test for object with first and second fields reversed and delete that instead.
-                weak_link_qs = WeakLinks.objects.filter(
-                        first_content_type_id = first_content_type.id,
-                        first_object_id = first_object_id,
-                        second_content_type_id = second_content_type.id,
-                        second_object_id = second_object_id_int
-                        )
-                if weak_link_qs:
-                    weak_link_instance = weak_link_qs[0]
-                    weak_link_instance.delete()
-                else:
+                if can_user_action:
+                    # transform request data to search for Weak Link obj to delete
+                    second_object_id_int = int(second_object_id)
+                    first_content_type = ContentType.objects.get(
+                            app_label='wildlifecompliance', 
+                            model=first_content_type_str)
+                    second_content_type = ContentType.objects.get(
+                            app_label='wildlifecompliance', 
+                            model=second_content_type_str)
+
+                    # If weak link obj not found, test for object with first and second fields reversed and delete that instead.
                     weak_link_qs = WeakLinks.objects.filter(
-                            first_content_type_id = second_content_type.id,
-                            first_object_id = second_object_id_int,
-                            second_content_type_id = first_content_type.id,
-                            second_object_id = first_object_id
+                            first_content_type_id = first_content_type.id,
+                            first_object_id = first_object_id,
+                            second_content_type_id = second_content_type.id,
+                            second_object_id = second_object_id_int
                             )
                     if weak_link_qs:
                         weak_link_instance = weak_link_qs[0]
                         weak_link_instance.delete()
+                    else:
+                        weak_link_qs = WeakLinks.objects.filter(
+                                first_content_type_id = second_content_type.id,
+                                first_object_id = second_object_id_int,
+                                second_content_type_id = first_content_type.id,
+                                second_object_id = first_object_id
+                                )
+                        if weak_link_qs:
+                            weak_link_instance = weak_link_qs[0]
+                            weak_link_instance.delete()
 
-                # derive parent (calling) object instance from ContentType
-                calling_instance = first_content_type.model_class().objects.get(id=first_object_id)
+                    # derive parent (calling) object instance from ContentType
+                    calling_instance = first_content_type.model_class().objects.get(id=first_object_id)
 
-                # get related items of calling_instance
-                related_items = get_related_items(calling_instance)
-                return Response(related_items)
+                    # get related items of calling_instance
+                    related_items = get_related_items(calling_instance)
+                    return Response(related_items)
+                else:
+                    content = {'message': 'User does not have permission to perform this action'}
+                    return Response(content, status=status.HTTP_403_FORBIDDEN)
         except Exception as e:
             print(traceback.print_exc())
             raise e
